Implemented error handling

+++ src/main/java/sk/ness/academy/service/CommentServiceImpl.java
@@ -2,10 +2,13 @@ package sk.ness.academy.service;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import sk.ness.academy.dao.ArticleDAO;
 import sk.ness.academy.domain.Article;
 import sk.ness.academy.domain.Comment;
+import sk.ness.academy.domain.ResourceNotFoundException;
 
 import javax.annotation.Resource;
 import javax.transaction.Transactional;
@@ -34,11 +37,22 @@ public class CommentServiceImpl implements CommentService{
     public List<Comment> readComments(Integer articleId) {
         Article article = this.articleDAO.findByID(articleId);
         List<Comment> comments = article.getComments();
+        if(comments.isEmpty()){
+            throw new ResourceNotFoundException("This article has no comments");
+        }
         return comments;
     }
 
-    public void deleteComment(Integer articleId, Integer commentId) {
+    public ResponseEntity deleteComment(Integer articleId, Integer commentId) {
         Session session = this.sessionFactory.getCurrentSession();
-        session.createSQLQuery("delete from comments where commentId = :commentId").setParameter("commentId", commentId).executeUpdate();
+        if (session.createSQLQuery("select * from comments where commentId = :commentId").setParameter("commentId", commentId).list().isEmpty()){
+        //   throw new ResourceNotFoundException("Comment with id " + commentId + " not exist");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("Comment with id " + commentId + " not exist");
+        } else {
+            session.createSQLQuery("delete from comments where commentId = :commentId").setParameter("commentId", commentId).executeUpdate();
+            return ResponseEntity.status(HttpStatus.CREATED)
+                    .body("Comment has been successfully deleted.");
+        }
     }
 }
